// async-race/src/enums/api.ts
export enum ApiAttributes {
    ID = 'id',
    WINS = 'wins',
    TYME = 'time',
    ASC = 'ASC',
    DESC = 'DESC',
    STARTED = 'started',
    STOPPED = 'stopped',
    DRIVE = 'drive',
}

export enum ApiMethods {
    POST = 'POST',
    DELETE = 'DELETE',
    PUT = 'PUT',
    PUTCH = 'PATCH',
}

export enum ApiPoints {
    GARAGE = 'garage',
    ENGINE = 'engine',
    WINNERS = 'winners',
}